Avoid sorting teams after creating raw pointers
Since raw pointers rely on the original order of the teams vec to be
preserved (bad design, to be sure). Now, a vector of references is
sorted instead and ranks assigned directly. Forces the return type for
the public teams function to be a bit odd, a vector of references
instead of a slice.